test(mail): align backend coverage with public APIs
